Remove bugged furnace warning
Fixes issue #19

This warning shows up every time you load a new world and it is a bug because even with other furnace recipes existing for, lets say, leather, or even if I add my own furnace recipes for it, it will still print this warning, therefore it is bugged and useless.

The warning is practically useless too since you can just check NEI furnace recipes for an item.